Create SSH - Potential Distributed Brute Force.kql
This analytic detects when a certain IP address has met a threshold of failed logon attempts to a number of hosts in a certain period of time. This may indicate an intrusion attempt against a particular ASN, subnet or company. Investigate the nature of the authentication attempts and take necessary action.